Reduce file count
