--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnAnchorBaseTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for NimbleTableColumn API on <see cref="NimbleTableColumnAnchor"/>
-/// </summary>
-public class NimbleTableColumnAnchorBaseTests : NimbleTableColumnTests<NimbleTableColumnAnchor>
-{
-}
