--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnAnchorFractionalWidthTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for FractionalWidthAPI on <see cref="NimbleTableColumnAnchor"/>
-/// </summary>
-public class NimbleTableColumnAnchorFractionalWidthTests : FractionalWidthBaseTests<NimbleTableColumnAnchor>
-{
-}
