--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnAnchorGroupableTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for GroupableAPI on <see cref="NimbleTableColumnAnchor"/>
-/// </summary>
-public class NimbleTableColumnAnchorGroupableTests : GroupableBaseTests<NimbleTableColumnAnchor>
-{
-}
