--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnTextBaseTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for NimbleTableColumn API on <see cref="NimbleTableColumnText"/>
-/// </summary>
-public class NimbleTableColumnTextBaseTests : NimbleTableColumnTests<NimbleTableColumnText>
-{
-}
