--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnTextFractionalWidthTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for FractionalWidthAPI on <see cref="NimbleTableColumnText"/>
-/// </summary>
-public class NimbleTableColumnTextFractionalWidthTests : FractionalWidthBaseTests<NimbleTableColumnText>
-{
-}
