--- packages/nimble-blazor/Tests/NimbleBlazor.Tests/Unit/Components/NimbleTableColumnTextGroupableTests.cs
@@ -1,9 +1,0 @@
-using Xunit;
-namespace NimbleBlazor.Tests.Unit.Components;
-
-/// <summary>
-/// Tests for GroupableAPI on <see cref="NimbleTableColumnText"/>
-/// </summary>
-public class NimbleTableColumnTextGroupableTests : GroupableBaseTests<NimbleTableColumnText>
-{
-}
